refactor: extract MorphTreeView base class from JsonLinesTreeView
