+++ src/App/Views/JsonLinesTreeView.cs
@@ -3,23 +3,19 @@ using System.Text.Json;
 using DataMorph.App.Views.JsonTreeNodes;
 using DataMorph.Engine.IO;
 using DataMorph.Engine.IO.JsonLines;
-using Terminal.Gui.Drivers;
-using Terminal.Gui.Input;
 using Terminal.Gui.Views;
 
 namespace DataMorph.App.Views;
 
 /// <summary>
-/// A TreeView that displays JSON Lines data as a hierarchical tree.
+/// A <see cref="MorphTreeView"/> that displays JSON Lines data as a hierarchical tree.
 /// Creates TreeNode instances from raw JSON bytes provided by the Engine layer.
 /// </summary>
-internal sealed class JsonLinesTreeView : TreeView
+internal sealed class JsonLinesTreeView : MorphTreeView
 {
     private const int InitialLoadCount = 50;
 
     private readonly RowByteCache _cache;
-    private readonly Action _onTableModeToggle;
-    private readonly VimKeyTranslator _vimKeys = new();
 
     /// <summary>
     /// Initializes a new instance of the <see cref="JsonLinesTreeView"/> class.
@@ -27,29 +23,10 @@ internal sealed class JsonLinesTreeView : TreeView
     /// <param name="indexer">The row indexer for the JSON Lines file.</param>
     /// <param name="onTableModeToggle">Callback invoked when the user presses 't'.</param>
     public JsonLinesTreeView(IRowIndexer indexer, Action onTableModeToggle)
+        : base(onTableModeToggle)
     {
         _cache = new RowByteCache(indexer);
-        _onTableModeToggle = onTableModeToggle;
         LoadInitialRootNodes();
-
-        Accepted += OnAccepted;
-    }
-
-    private void OnAccepted(object? sender, CommandEventArgs e)
-    {
-        var node = SelectedObject;
-        if (node is null)
-        {
-            return;
-        }
-
-        if (IsExpanded(node))
-        {
-            Collapse(node);
-            return;
-        }
-
-        Expand(node);
     }
 
     private void LoadInitialRootNodes()
@@ -130,58 +107,6 @@ internal sealed class JsonLinesTreeView : TreeView
             JsonTokenType.Null => JsonValueKind.Null,
             _ => JsonValueKind.Undefined,
         };
-    }
-
-    /// <inheritdoc/>
-    protected override bool OnKeyDown(Key key)
-    {
-        if (key.KeyCode == KeyCode.T)
-        {
-            _onTableModeToggle();
-            return true;
-        }
-
-        var action = _vimKeys.Translate(key.KeyCode);
-
-        return action switch
-        {
-            VimAction.PendingGSequence => true,
-            VimAction.MoveDown => ConsumeAction(() =>
-                AdjustSelection(offset: 1, expandSelection: false)
-            ),
-            VimAction.MoveUp => ConsumeAction(() =>
-                AdjustSelection(offset: -1, expandSelection: false)
-            ),
-            VimAction.PageDown => ConsumeAction(() =>
-                AdjustSelection(offset: Viewport.Height, expandSelection: false)
-            ),
-            VimAction.PageUp => ConsumeAction(() =>
-                AdjustSelection(offset: -Viewport.Height, expandSelection: false)
-            ),
-            VimAction.MoveLeft => base.OnKeyDown(new Key(KeyCode.CursorLeft)),
-            VimAction.MoveRight => base.OnKeyDown(new Key(KeyCode.CursorRight)),
-            VimAction.GoToFirst => ConsumeAction(GoToFirst),
-            VimAction.GoToEnd => ConsumeAction(GoToEnd),
-            _ => HandleNonVimKey(key),
-        };
-    }
-
-    private bool HandleNonVimKey(Key key)
-    {
-        // Prevent global shortcut keys from being consumed by TreeView's incremental search.
-        // By returning false, we let these keys bubble up to AppKeyHandler.
-        if (AppKeyHandler.IsGlobalShortcut(key.KeyCode))
-        {
-            return false;
-        }
-
-        return base.OnKeyDown(key);
-    }
-
-    private static bool ConsumeAction(Action action)
-    {
-        action();
-        return true;
     }
 
     /// <inheritdoc/>
